fix: resolve run-seal windows portability and sonarcloud findings
