+++ src/aptl/cli/runs.py
@@ -23,6 +23,48 @@ def _get_store(project_dir: Path) -> LocalRunStore:
     return resolve_run_store(project_dir)
 
 
+def _print_run_table(store: LocalRunStore, run_ids: list[str]) -> None:
+    """Render the runs as a rich table (interactive TTY output)."""
+    table = Table(title="Experiment Runs")
+    table.add_column("Run ID", style="cyan", no_wrap=True)
+    table.add_column("Scenario")
+    table.add_column("Started")
+    table.add_column("Duration")
+    table.add_column("Flags")
+
+    for run_id in run_ids:
+        try:
+            summary = summarize_manifest(store.get_run_manifest(run_id))
+            duration = summary.duration_seconds or 0
+            minutes = int(duration // 60)
+            seconds = int(duration % 60)
+            flags = "0" if summary.flags_captured is None else str(summary.flags_captured)
+            table.add_row(
+                run_id[:12] + "...",
+                summary.scenario_name,
+                summary.started_at[:19],
+                f"{minutes}m {seconds}s",
+                flags,
+            )
+        except (FileNotFoundError, KeyError) as e:
+            log.warning("Skipping run %s: %s", run_id, e)
+            table.add_row(run_id[:12] + "...", "[red]ERROR[/red]", "", "", "")
+
+    console.print(table)
+
+
+def _print_run_lines(store: LocalRunStore, run_ids: list[str]) -> None:
+    """Render the runs as plain tab-separated lines (non-TTY output)."""
+    for run_id in run_ids:
+        try:
+            summary = summarize_manifest(store.get_run_manifest(run_id))
+            typer.echo(
+                f"{run_id}\t{summary.scenario_name}\t{summary.started_at[:19]}"
+            )
+        except (FileNotFoundError, KeyError):
+            typer.echo(f"{run_id}\tERROR")
+
+
 @app.command("list")
 def list_runs(
     project_dir: Path = typer.Option(
@@ -50,41 +92,9 @@ def list_runs(
     run_ids = list(reversed(run_ids))[:limit]
 
     if sys.stdout.isatty():
-        table = Table(title="Experiment Runs")
-        table.add_column("Run ID", style="cyan", no_wrap=True)
-        table.add_column("Scenario")
-        table.add_column("Started")
-        table.add_column("Duration")
-        table.add_column("Flags")
-
-        for run_id in run_ids:
-            try:
-                summary = summarize_manifest(store.get_run_manifest(run_id))
-                duration = summary.duration_seconds or 0
-                minutes = int(duration // 60)
-                seconds = int(duration % 60)
-                flags = "0" if summary.flags_captured is None else str(summary.flags_captured)
-                table.add_row(
-                    run_id[:12] + "...",
-                    summary.scenario_name,
-                    summary.started_at[:19],
-                    f"{minutes}m {seconds}s",
-                    flags,
-                )
-            except (FileNotFoundError, KeyError) as e:
-                log.warning("Skipping run %s: %s", run_id, e)
-                table.add_row(run_id[:12] + "...", "[red]ERROR[/red]", "", "", "")
-
-        console.print(table)
+        _print_run_table(store, run_ids)
     else:
-        for run_id in run_ids:
-            try:
-                summary = summarize_manifest(store.get_run_manifest(run_id))
-                typer.echo(
-                    f"{run_id}\t{summary.scenario_name}\t{summary.started_at[:19]}"
-                )
-            except (FileNotFoundError, KeyError):
-                typer.echo(f"{run_id}\tERROR")
+        _print_run_lines(store, run_ids)
 
     typer.echo(f"\n{len(run_ids)} run(s) shown (of {len(store.list_runs())} total)")
 
@@ -101,18 +111,7 @@ def show_run(
 ) -> None:
     """Show details of a specific run."""
     store = _get_store(project_dir)
-
-    # Support prefix matching
-    all_runs = store.list_runs()
-    matches = [r for r in all_runs if r.startswith(run_id)]
-    if not matches:
-        typer.echo(f"No run found matching '{run_id}'")
-        raise typer.Exit(code=1)
-    if len(matches) > 1:
-        typer.echo(f"Ambiguous run ID '{run_id}', matches: {', '.join(matches[:5])}")
-        raise typer.Exit(code=1)
-
-    resolved_id = matches[0]
+    resolved_id = _resolve_run_id(store, run_id)
 
     try:
         summary = summarize_manifest(store.get_run_manifest(resolved_id))
@@ -137,22 +136,7 @@ def show_run(
     if summary.containers:
         typer.echo(f"  Containers:   {', '.join(summary.containers)}")
 
-    # List files in run directory
-    run_path = store.get_run_path(resolved_id)
-    if run_path.exists():
-        typer.echo("")
-        typer.echo("Files:")
-        for item in sorted(run_path.rglob("*")):
-            if item.is_file():
-                rel = item.relative_to(run_path)
-                size = item.stat().st_size
-                if size > 1024 * 1024:
-                    size_str = f"{size / 1024 / 1024:.1f} MB"
-                elif size > 1024:
-                    size_str = f"{size / 1024:.1f} KB"
-                else:
-                    size_str = f"{size} B"
-                typer.echo(f"  {rel}  ({size_str})")
+    _echo_run_files(store.get_run_path(resolved_id))
 
 
 @app.command("path")
@@ -178,6 +162,27 @@ def run_path(
         raise typer.Exit(code=1)
 
     typer.echo(str(store.get_run_path(matches[0])))
+
+
+def _format_size(size: int) -> str:
+    """Render a byte count as a human-readable size string."""
+    if size > 1024 * 1024:
+        return f"{size / 1024 / 1024:.1f} MB"
+    if size > 1024:
+        return f"{size / 1024:.1f} KB"
+    return f"{size} B"
+
+
+def _echo_run_files(run_path: Path) -> None:
+    """List the files under a run directory with their sizes."""
+    if not run_path.exists():
+        return
+    typer.echo("")
+    typer.echo("Files:")
+    for item in sorted(run_path.rglob("*")):
+        if item.is_file():
+            rel = item.relative_to(run_path)
+            typer.echo(f"  {rel}  ({_format_size(item.stat().st_size)})")
 
 
 def _resolve_run_id(store: LocalRunStore, run_id: str) -> str:
